+++ app/templates/routes/blog/timeline.html
@@ -92,9 +92,9 @@
 
 <!-- Confirm Popup-->
 <div id="delete-modal" class="fixed inset-0 bg-primary bg-opacity-50 flex items-center justify-center hidden">
-  <div class="flex flex-col justify-center items-center bg-secondary w-1/3 h-1/3 rounded-lg shadow-lg border-2 border-dashed">
-    <h2 class="text-2xl font-bold mb-4">Confirm Deletion</h2>
-    <p id="post-details" class="mb-2"></p>
+  <div class="flex flex-col justify-center items-center bg-secondary w-1/3 rounded-lg shadow-lg border-2 border-dashed">
+    <h2 class="text-2xl font-bold mb-4 mt-6">Confirm Deletion</h2>
+    <p id="post-details" class="mb-2 mx-3 text-center"></p>
     <p class="mb-6">Are you sure you want to delete this post?</p>
     
 
@@ -117,9 +117,9 @@
       />
     </form>
 
-    <p id="failure"></p>
+    <p id="failure" class="text-primary">test test test</p>
 
-    <div class="mt-5 flex justify-center">
+    <div class="mt-5 flex justify-center mb-10">
       <button id="cancel-button" class="bg-gray-300 hover:bg-gray-400 text-gray-800 font-bold py-2 px-4 rounded mr-2">Cancel</button>
       <button id="confirm-delete-button" class="bg-red-500 hover:bg-red-600 text-white font-bold py-2 px-4 rounded">Delete</button>
     </div>
@@ -235,7 +235,7 @@
       const formData = new FormData();
       formData.append("name", name);
       formData.append("start", start);
-      formData.append("end", start);
+      formData.append("end", new Date(new Date(start).getTime() + 1000).toUTCString());
       formData.append("password", password)
 
       const response = await fetch(apiUrl, {
@@ -243,8 +243,9 @@
         body: formData,
       });
 
-      return true;
 
+      const result = await response.json(); // Parse the JSON response
+      return result; // Return the parsed result
       if (!response.ok) {
         throw new Error("API Server Error " + response.statusText);
         return false;
@@ -268,31 +269,37 @@
 
     // Show modal
     const deleteModal = document.getElementById("delete-modal");
-    document.getElementById("post-details").innerHTML = `${name} By: ${email} </p>`
+    document.getElementById("post-details").innerHTML = `${name} <br/>By: ${email} </p>`
     deleteModal.classList.remove("hidden");
 
 
-    const processDelete = async () => {
+    const processDelete = async () => {      
       // Grabbing Confirm and Password
       const confirm = document.getElementById('delete-confirm').value;
       const password = document.getElementById('delete-password').value;
 
       
-      if (confirm == "confirm") {   // Confirm delete button event listener
+      if (confirm === "confirm") {   // Confirm delete button event listener
+        const confirmButton = document.getElementById("confirm-delete-button");
+        confirmButton.disabled = true;
         document.getElementById("confirm-delete-button").onclick = 
           async () => {
-            if (await deletePosts(name, start_date, password)) { // Correct Password
+            const response = await deletePosts(name, start_date, password)
+            console.log(response)
+            if (response.result === "Success") { // Correct Password
               deleteModal.classList.add("hidden");
               await renderPosts();
 
               // Clearing Out Error Messages
-              document.getElementById("failure").className = "";
-              document.getElementById("failure").innerText = "";
+              document.getElementById("failure").className = "text-primary";
+              document.getElementById("failure").innerText = "test test test";
             }else{  // Incorrect Password
+              console.log('incorrect')
               document.getElementById("failure").className = "italic text-red-300 font-thin";
               document.getElementById("failure").innerText = "password is incorrect";
             }
         };
+        confirmButton.disabled = false;
       }else{  // Non Confirm
         document.getElementById("failure").className = "italic text-white font-thin";
         document.getElementById("failure").innerText = "please type confirm to confirm deletion";
@@ -301,12 +308,13 @@
 
     // EVENT LISTENER
     document.getElementById("cancel-button").onclick = () => { 
-      document.getElementById("failure").className = "";
-      document.getElementById("failure").innerText = "";
+      document.getElementById("failure").className = "text-primary";
+      document.getElementById("failure").innerText = "test";
       deleteModal.classList.add("hidden"); 
     };
     document.getElementById("confirm-delete-button").onclick = processDelete;
   };
+
 
   // EVENT LISTENERS
   document.addEventListener("DOMContentLoaded", renderPosts);
@@ -322,6 +330,12 @@
       await renderDelete(event);
     }
   });
+  document.getElementById("cancel-button").onclick = () => { 
+    document.getElementById("failure").className = "text-primary";
+    document.getElementById("failure").innerText = "test";
+    deleteModal.classList.add("hidden"); 
+  };
+  document.getElementById("confirm-delete-button").onclick = processDelete;
 
 </script>
 {% endblock %}
