+++ docker-compose-prod.yml
@@ -1,0 +1,24 @@
+version: "3"
+services:
+  myportfolio:
+    container_name: MLH-portfolio-website
+    build: .
+    restart: always
+    env_file:
+      - .env
+    ports:
+      - "3004:3004"
+    depends_on:
+      - mysql
+  
+  mysql:
+    container_name: mysql
+    image: mariadb
+    restart: always
+    env_file:
+      - .env
+    volumes:
+      - mydatabase:/var/lib/mysql
+
+volumes:
+  mydatabase:
