+++ docker-compose.yml
@@ -1,0 +1,28 @@
+version: "3"
+services:
+  myportfolio:
+    container_name: MLH-portfolio-website
+    build: .
+    restart: always
+    env_file:
+      - .env
+    environment:
+      - FLASK_ENV=development
+    ports:
+      - "3004:3004"
+    volumes:
+      - .:/myportfolio
+    depends_on:
+      - mysql
+  
+  mysql:
+    container_name: mysql
+    image: mariadb
+    restart: always
+    env_file:
+      - .env
+    volumes:
+      - mydatabase:/var/lib/mysql
+
+volumes:
+  mydatabase:
